+++ include/Graph.hpp
@@ -12,7 +12,7 @@ namespace Graph
     struct WeightedGraphEdge
     {
         int to{};
-        float weight{1.0f};
+        double weight{1.0f};
         bool isClosed{false};
 
         void toggleClosed();
@@ -28,28 +28,28 @@ namespace Graph
 
         void add_vertex(int const id);
 
-        void add_directed_edge(int const from, int const to, float const weight = 1.0f);
+        void add_directed_edge(int const from, int const to, double const weight = 1.0f);
 
         bool operator==(WeightedGraph const &other) const;
         bool operator!=(WeightedGraph const &other) const;
 
         void print_BFS(int const start) const;
 
-        unordered_map<int, pair<float, int>> dijkstra(int const &start, int const &end);
+        unordered_map<int, pair<double, int>> dijkstra(int const &start, int const &end);
     };
 
-    WeightedGraph build_from_adjacency_matrix(vector<vector<float>> const &adjacency_matrix);
-    void display_shortest_path(unordered_map<int, pair<float, int>> tab, int start, int end);
+    WeightedGraph build_from_adjacency_matrix(vector<vector<double>> const &adjacency_matrix);
+    void display_shortest_path(unordered_map<int, pair<double, int>> tab, int start, int end);
 
 }
 
 ostream &operator<<(ostream &os, const Graph::WeightedGraph graph);
-ostream &operator<<(ostream &os, const vector<vector<float>> adj_mat);
-ostream &operator<<(ostream &os, const unordered_map<int, pair<float, int>> map);
+ostream &operator<<(ostream &os, const vector<vector<double>> adj_mat);
+ostream &operator<<(ostream &os, const unordered_map<int, pair<double, int>> map);
 bool isIn(const vector<int> &vec, int val);
 
 struct AdjacencyMatrix
 {
-    vector<vector<float>> matrix;
+    vector<vector<double>> matrix;
     void setMatrix();
 };
